+++ src/aks-preview/HISTORY.rst
@@ -12,6 +12,11 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 
+18.0.0b11
++++++++
+* Vendor new SDK and bump API version to 2025-04-02-preview.
+* Deprecate options `--safeguards-level`, `--safeguards-version` and `--safeguards-excluded-ns` in `az aks create` and `az aks update` commands.
+
 18.0.0b10
 +++++++
 * Wrap the ARG call in the managed namespace list command